+++ src/App.js
@@ -3,16 +3,20 @@ import { Login } from "./components/auth/Login"
 import { Register } from "./components/auth/Register"
 import { NavBar } from "./components/nav/NavBar"
 import { ApplicationViews } from "./components/views/ApplicationViews"
+import { Authorized } from "./components/views/Authorized"
 
 export const App = () => {
     return <Routes>
         <Route path="login" element={<Login/>}/>
         <Route path="register" element={<Register/>}/>
         <Route path="*" element={
-            <>
-                <NavBar/>
-                <ApplicationViews/>
-            </>
+            <Authorized>
+                <>
+                    <NavBar/>
+                    <ApplicationViews/>
+                </>
+            </Authorized>
+            
         }/>
     </Routes>
 }
