+++ src/components/nav/NavBar.js
@@ -1,16 +1,20 @@
-// If a current user exists, display the logged in nav bar
-// Else display the logged out nav bar
+// User can click an account icon in the nav bar that will display a list of route options
+// User can click on a logout button
 
-import { LoggedInNav } from "./LoggedInNav"
-import { LoggedOutNav } from "./LoggedOutNav"
+import { useEffect, useState } from "react"
+import { useNavigate } from "react-router-dom"
 
 export const NavBar = () => {
 
-    const localUser = localStorage.getItem("current_user")
+    const navigate = useNavigate()
 
-    if (localUser) {
-        return <LoggedInNav/>
-    } else {
-        return <LoggedOutNav/>
-    }
+    return (
+        <>
+            <button onClick={() => navigate("/account")} className="navbar__button navbar__account">Account</button>
+            <button onClick={() => {
+                localStorage.removeItem("current_user")
+                navigate("/", {replace: true})
+                }} className="navbar__button navbar__logout">Logout</button>
+        </>
+    )
 }
